requirements.txt
africastalking==1.2.5
APScheduler==3.10.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contextlib2==21.6.0
cryptography==41.0.1
cssselect2==0.7.0
Django==3.2
django-apscheduler==0.6.2
django-rest-framework==0.1.0
django-tenants==3.5.0
djangorestframework==3.14.0
docx2txt==0.8
fuzzywuzzy==0.18.0
html5lib==1.1
idna==3.4
install==1.3.5
joblib==1.2.0
lxml==4.9.2
nltk==3.8.1
numpy==1.24.3
opencv-python==4.7.0.72
oscrypto==1.3.0
pandas==2.0.2
Pillow==9.5.0
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.10.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
regex==2023.6.3
reportlab==3.6.13
requests==2.31.0
schema==0.7.5
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sklearn3==0.0.1
sqlparse==0.4.4
svglib==1.5.1
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
webencodings==0.5.1
xhtml2pdf==0.2.11
